Admin can add sensor to Server. Device database use MongoDB Sensor dataBase use Mongodb

+++ src/main/java/org/maana/iot_project/services/AdminDeviceManagementService.java
@@ -23,7 +23,7 @@ public class AdminDeviceManagementService {
 		return dataBase.getDevice(deviceId);
 	}
 
-	private int noOfDevices() {
+	private long noOfDevices() {
 
 		return dataBase.noOfDevices();
 	}
